Extract alert methods to separate module
