+++ lib/onlyoffice_webdriver_wrapper/webdriver.rb
@@ -6,6 +6,7 @@ require 'selenium-webdriver'
 require 'uri'
 require_relative 'helpers/chrome_helper'
 require_relative 'helpers/firefox_helper'
+require_relative 'webdriver/webdriver_alert_helper'
 require_relative 'webdriver/webdriver_attributes_helper'
 require_relative 'webdriver/webdriver_type_helper'
 require_relative 'webdriver/webdriver_exceptions'
@@ -19,6 +20,7 @@ module OnlyofficeWebdriverWrapper
     include ChromeHelper
     include FirefoxHelper
     include RubyHelper
+    include WebdriverAlertHelper
     include WebdriverAttributesHelper
     include WebdriverTypeHelper
     include WebdriverHelper
@@ -108,30 +110,6 @@ module OnlyofficeWebdriverWrapper
       end
     rescue
       nil
-    end
-
-    def alert_confirm
-      return if @browser == :ie
-      begin
-        @driver.switch_to.alert.accept
-      rescue
-        Selenium::WebDriver::Error::NoAlertOpenError
-      end
-    end
-
-    # Check if alert exists
-    # @return [True, false]
-    def alert_exists?
-      @driver.switch_to.alert.text
-      true
-    rescue Selenium::WebDriver::Error::NoAlertOpenError
-      false
-    end
-
-    # Get alert text
-    # @return [String] text inside alert
-    def alert_text
-      @driver.switch_to.alert.text
     end
 
     def set_text_to_iframe(element, text)
